Main (Stable Version 1.0)

+++ src/main/java/com/Revature/eCommerce/models/History.java
@@ -16,5 +16,5 @@ public class History
 {
     private String id;
     private String userId;
-    private float totalCost;
+    private int totalCost;
 }
